refactor(spec): keep portable imports in the package builder
